Fix the programmer dialogs
Overhaul of the programmer dialogs and the way the programmers work.
Now the programmers do not hold state in between button pushes and
do not hold state by just being open. This means you can still
interact with the target while the programmer dialogs are up. This was
not possible before, and would create really confusing errors.

Pin states are now saved before the programming begins and restored when
the programming is complete. This boot pins/reset pins are not left
around after the process is finished.

updateUI is called during programming for STM32 because it takes longer
so the UI does not become non-responsive for too long