Fix for no target disconnect for scope disconnect
The target used to stay connected when the scope was unplugged.
This does not make sense for SimpleSerial target because it keeps
a reference to the old target around and this creates weird errors
later on. This is now fixed with a signal that is emitted when the
scope is disconnected. The signal is connected to the scope's target
instance. Now when unplugging the scope the target connected to the
scope is also disconnected.