Fix for self.scope access in Capture GUI if None
The changes make the error messages shown in the GUI when accesssing
the self.scope attribute, which intially is None, more understandable.
Before the error was raises later when you tried to access the attribute
of the NoneType returned. The access of the self.scope attribute of GUI
now checks if it is None and returns a NoneTypeScope instance. If any
of the attributes of the NoneTypeScope are accessed it will raise a
error hinting that the scope is not connected. The same was dons for the
self.target attribute of the GUI.