Add a read timeout to simple serial.
The simple serial target reads data from the serial. By default it does
this using a timeout of 500ms and this is not always suitable. Secondly
the read serial target does not use a timeout when forwarding the data
to the glitch module.

This commit adds a read_timeout variable to SimpleSerial to make it
possible to change it. It also introduces the same timeout in when using
the glitch module.