Fix "langid" for multiple attached devices.
When there are multiple attached devices and the user does
not have access to some of them the "langid" error will
be raised by each device that the user does not have
permissions for. It now warns the user but keeps looking
for any device which it has permissions to access. This
behaviouy is limited to when there is more than one device
attached.