Add close method for Project along with many tests for project.
The project importing, exporting, closing, removing, saving was used to
test if the close method works as intended. project.remove uses the
close method to clean up before removing. The fix was for crazy stuff
when trying to remove a project. It would raise file in use errors.
This is now fixed.