Add scrollbar for fixed sidebar when content overflows.
There is a weird issue when setting max-height: 100% that cause the
scrollable section to cut off the bottom part of the content. Almost
like the estimate for 100% height is off. I do not know enough to
properly fix it so I set the max-height to 93 %. Also is bottom: 0 is
set the scrollbar works again but the sidebar is always stuck to the
bottom of the page. This was the only other thing I explored.