Expose data type arguments for capturing non-float waves
Both the OpenADC scope and the Trace class can already output and store
integers instead of doubles for the waves, respectively. But the
corresponding arguments were not exposed to the API previously. This
commit exposes these arguments to the API to allow users capture e.g.
waves as uint16 instead of doubles. This allows to reduce the memory
and storage requirements of long captures by roughly 4x.

This is related to NewAETech/ChipWhisperer#344.