Husky: fix how LA/swo trace clock is set.
Previously avoided using MMCM's secondary divider because it often
couldn't be set. Turns out this happens when the MMCM VCO is taken out
of range. Add instructions to avoid this and give a warning if it
happens. With this fix, there is now greater flexibility in setting
clock frequencies, and we ensure that the MMCM is operating within its
allowed range.